refactor: extract the partition handling from the consumer
